Alterando tamanho do cpf
Alterando CPF para suportar Long int

+++ db/mysql 5.7+/biblioteca.sql
@@ -6,7 +6,7 @@ USE biblioteca;
 
 DROP TABLE IF EXISTS `alunos`;
 CREATE TABLE IF NOT EXISTS `alunos` (
-  `id` INT NOT NULL,
+  `id` BIGINT NOT NULL,
   `nome` VARCHAR(255) NOT NULL,
   `senha` VARCHAR(255) NOT NULL,
   `email` VARCHAR(255) NOT NULL UNIQUE,
